Move revpi-device-info into package
There is no need to maintain a separate script for revpi-device-info.
Therefore the script is moved into the package's main method. This will
allow to call the library like this 'python3 -m revpi_device_info' and
also can be used further as a console_script with a proper setup.py

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>